Final cleanups
