+++ src/Spiral/SpinWaveTheorySpiral.jl
@@ -21,9 +21,12 @@ struct SpinWaveTheorySpiral <: AbstractSpinWaveTheory
     swt :: SpinWaveTheory
     k :: Vec3
     axis :: Vec3
+    buffers :: Vector{Array{CMat3, 2}}
 
     function SpinWaveTheorySpiral(sys::System; k::AbstractVector, axis::AbstractVector, measure::Union{Nothing, MeasureSpec}, regularization=1e-8)
-        return new(SpinWaveTheory(sys; measure, regularization), k, axis)
+        L = length(eachsite(sys))
+        buffers = [zeros(CMat3, L, L) for _ in 1:6]
+        return new(SpinWaveTheory(sys; measure, regularization), k, axis, buffers)
     end
 end
 
@@ -34,6 +37,43 @@ function construct_uniaxial_anisotropy(; axis, c20=0., c40=0., c60=0., s)
     # Rotate operator into global frame, defined by axis
     R = rotation_between_vectors(axis, [0, 0, 1])
     return rotate_operator(op, R)
+end
+
+# Sunny's original implementation followed Toth and Lake, which employed
+# pairwise exchange interactions J in real-space. Fourier-space exchange J(q) is
+# required, for example, to support long-range dipole-dipole interactions. Alin
+# Niraula and Xiaojian Bai developed and implemented the approprate
+# generalization in https://github.com/SunnySuite/Sunny.jl/pull/304.
+
+function fourier_bilinear_interaction!(J_k, swt::SpinWaveTheory, q)
+    (; sys, data) = swt
+    (; local_rotations) = data
+    (; gs) = sys
+    @assert sys.dims == (1, 1, 1) "System must have only a single cell"
+    Rs = local_rotations
+    Na = natoms(sys.crystal)
+    fill!(J_k, zero(CMat3))
+
+    for i in 1:natoms(sys.crystal)
+        for coupling in sys.interactions_union[i].pair
+            (; isculled, bond, bilin) = coupling
+            isculled && break
+
+            (; j, n) = bond
+            J_lab = Rs[i] * Mat3(bilin*I) * Rs[j]' # Undo transformation in `swt_data`
+            J = exp(-2π * im * dot(q, n)) * J_lab
+            J_k[i, j] += J
+            J_k[j, i] += J'
+        end
+    end
+
+    if !isnothing(sys.ewald)
+        A = precompute_dipole_ewald_at_wavevector(sys.crystal, (1,1,1), -q) * sys.ewald.μ0_μB²
+        A = reshape(A, Na, Na)
+        for i in 1:Na, j in 1:Na
+            J_k[i, j] += gs[i]' * A[i, j] * gs[j]
+        end
+    end
 end
 
 
@@ -47,42 +87,73 @@ function swt_hamiltonian_dipole_spiral!(H::Matrix{ComplexF64}, sswt::SpinWaveThe
     @assert size(H) == (2L, 2L)
     H .= 0.0
 
+    # preallocation
+    H11 = view(H, 1:L, 1:L)
+    H12 = view(H, 1:L, L+1:2L)
+    H21 = view(H, L+1:2L, 1:L)
+    H22 = view(H, L+1:2L, L+1:2L)
+    nx = SA[0 -axis[3] axis[2]; axis[3] 0 -axis[1]; -axis[2] axis[1] 0]
+    R2 = axis * axis'
+    R1 = (I - im * nx - R2) / 2
+    Rs = local_rotations
+
+    q_reshaped = q_reshaped + (branch - 2) .* k
+
     # Add pairwise bilinear term
-    for ints in sys.interactions_union
 
-        for c in ints.pair
-            (; i, j, n) = c.bond
-            θ = (2*π * dot(k,n))
-            Rn = axis_angle_to_matrix(axis, θ)
+    Jq, Jqmk, Jqpk, J0k, J0mk, J0pk = sswt.buffers
+    fourier_bilinear_interaction!(Jq,   swt, q_reshaped)
+    fourier_bilinear_interaction!(Jqmk, swt, q_reshaped .- k)
+    fourier_bilinear_interaction!(Jqpk, swt, q_reshaped .+ k)
+    fourier_bilinear_interaction!(J0k,  swt, zero(Vec3))
+    fourier_bilinear_interaction!(J0mk, swt, -k)
+    fourier_bilinear_interaction!(J0pk, swt, +k)
 
-            # Undo rotations that were created in SpinWaveTheory.jl
-            Ri = local_rotations[i]
-            Rj = local_rotations[j]
-            J = Ri * c.bilin * Rj'
-
-            Jij = (J * Rn + Rn * J) ./ 2
-            phase = exp(2π * im * dot(q_reshaped + (branch-2)*k, n))
-
-            sj = sqrtS[j]^2
-            sij = sqrtS[i] * sqrtS[j]
-
-            ui = Ri[:,1] + im*Ri[:,2]
-            uj = Rj[:,1] + im*Rj[:,2]
-            vi = Ri[:,3]
-            vj = Rj[:,3]
-
-            H[i,j]     += (sij/2) * (transpose(ui)) * Jij * conj(uj) * phase
-            H[i+L,j+L] += (sij/2) * conj((transpose(ui)) * Jij * conj(uj)) * phase
-
-            H[i,j+L]   += (sij/2) * (transpose(ui) * Jij * uj) * phase
-            H[j+L,i]   += (sij/2) * conj(transpose(ui) * Jij * uj * phase)
-
-            H[i,i]     -= sj * transpose(vi) * Jij * vj
-            H[i+L,i+L] -= sj * transpose(vi) * Jij * vj
-
-            iszero(c.biquad) || error("Biquadratic interactions not supported")
+    for i in 1:L, j in 1:L
+        Jq1   = Jq[i, j]
+        Jqmk1 = Jqmk[i, j]
+        Jqpk1 = Jqpk[i, j]
+        J0k1  = J0k[i, j]
+        J0pk1 = J0pk[i, j]
+        J0mk1 = J0mk[i, j]
+        case = spiral_propagation_case(k)
+        if case == 1
+            J = Jq1
+            J0 = J0k1
+        elseif case == 2
+            J = R2 * Jq1* R2 + conj(R1) * Jqpk1 * conj(R1) + R1 * Jqmk1 * R1 + R1 * Jqpk1 * conj(R1) + conj(R1) * Jqmk1 * R1
+            J0 = R2 * J0k1 * R2 + conj(R1) * J0pk1 * conj(R1) + R1 * J0mk1 * R1 + R1 * J0pk1 * conj(R1) + conj(R1) * J0mk1 * R1
+        else @assert case == 3
+            J = R2 * Jq1* R2 + conj(R1) * Jqpk1 * conj(R1) + R1 * Jqmk1 * R1
+            J0 = R2 * J0k1 * R2 + conj(R1) * J0pk1 * conj(R1) + R1 * J0mk1 * R1
         end
-    end
+
+        # Perform same transformation as appears in usual bilinear exchange.
+        # Rⱼ denotes a rotation from ẑ to the ground state dipole Sⱼ.
+        J = sqrtS[i] * sqrtS[j] * Rs[i]' * J * Rs[j]
+        J0 = sqrtS[i] * sqrtS[j] * Rs[i]' * J0 * Rs[j]
+
+        # Interactions for Jˣˣ, Jʸʸ, Jˣʸ, and Jʸˣ at wavevector q.
+        Q⁺ = 0.25 * (J[1, 1] + J[2, 2] - im*(J[1, 2] - J[2, 1]))
+        Q⁻ = 0.25 * (J[1, 1] + J[2, 2] + im*(J[1, 2] - J[2, 1]))
+        H11[i, j] += conj(Q⁻)
+        H11[j, i] += Q⁻
+        H22[i, j] += conj(Q⁺)
+        H22[j, i] += Q⁺
+
+        P⁺ = 0.25 * (J[1, 1] - J[2, 2] - im*(J[1, 2] + J[2, 1]))
+        P⁻ = 0.25 * (J[1, 1] - J[2, 2] + im*(J[1, 2] + J[2, 1]))
+        H21[i, j] += conj(P⁻)
+        H21[j, i] += P⁺
+        H12[i, j] += conj(P⁺)
+        H12[j, i] += P⁻
+
+        # Interactions for Jᶻᶻ at wavevector (0,0,0).
+        H11[i, i] -= 0.5 * J0[3, 3]
+        H11[j, j] -= 0.5 * J0[3, 3]
+        H22[i, i] -= 0.5 * J0[3, 3]
+        H22[j, j] -= 0.5 * J0[3, 3]
+     end
 
     # Add Zeeman term
     for i in 1:L
@@ -101,8 +172,6 @@ function swt_hamiltonian_dipole_spiral!(H::Matrix{ComplexF64}, sswt::SpinWaveThe
         H[i, i+L]   += 2s*(c2[1]+im*c2[5]) + 12s^3*(c4[3]+im*c4[7]) + 32s^5*(c6[5]+im*c6[9])
         H[i+L, i]   += 2s*(c2[1]-im*c2[5]) + 12s^3*(c4[3]-im*c4[7]) + 32s^5*(c6[5]-im*c6[9])
     end
-
-    isnothing(sys.ewald) || error("Ewald interactions not yet supported")
 
     @assert diffnorm2(H, H') < 1e-12
     hermitianpart!(H)
@@ -175,15 +244,16 @@ end
 
 
 function intensities_bands(sswt::SpinWaveTheorySpiral, qpts; kT=0) # TODO: branch=nothing
-    (; swt, axis) = sswt
+    (; swt, k, axis) = sswt
     (; sys, data, measure) = swt
+    (; local_rotations, observables_localized, sqrtS) = data
+
     isempty(measure.observables) && error("No observables! Construct SpinWaveTheorySpiral with a `measure` argument.")
     sys.mode == :SUN && error("SU(N) mode not supported for spiral calculation")
     @assert sys.mode in (:dipole, :dipole_uncorrected)
 
     qpts = convert(AbstractQPoints, qpts)
     cryst = orig_crystal(sys)
-    R = data.local_rotations
 
     # Number of atoms in magnetic cell
     @assert sys.dims == (1,1,1)
@@ -196,7 +266,6 @@ function intensities_bands(sswt::SpinWaveTheorySpiral, qpts; kT=0) # TODO: branc
     Nq = length(qpts.qs)
 
     # Rotation matrices associated with `axis`
-    CMat3 = SMatrix{3, 3, ComplexF64, 9}
     nx = CMat3([0 -axis[3] axis[2]; axis[3] 0 -axis[1]; -axis[2] axis[1] 0])
     R2 = CMat3(axis * axis')
     R1 = (1/2) .* CMat3(I - im .* nx - R2)
@@ -205,71 +274,66 @@ function intensities_bands(sswt::SpinWaveTheorySpiral, qpts; kT=0) # TODO: branc
     H = zeros(ComplexF64, 2L, 2L)
     T0 = zeros(ComplexF64, 2L, 2L)
     T = zeros(ComplexF64, 2L, 2L, 3)
-    YZVW = zeros(ComplexF64, 2L, 2L, 3, 3) # [[Y Z]; [V W]]
-    Y = view(YZVW, 1:L, 1:L, :, :)
-    Z = view(YZVW, 1:L, L+1:2L, :, :)
-    V = view(YZVW, L+1:2L, 1:L, :, :)
-    W = view(YZVW, L+1:2L, L+1:2L, :, :)
-    
     disp = zeros(Float64, L, 3, Nq)
     intensity = zeros(eltype(measure), L, 3, Nq)
     disp_flat = reshape(disp, 3L, Nq)
     intensity_flat = reshape(intensity, 3L, Nq)
-    S = zeros(ComplexF64, 3, 3, L, 3)
-
-    # Like Avec_pref
-    c = zeros(ComplexF64, Na)
+    S = zeros(CMat3, L, 3)
+    Avec_pref = zeros(ComplexF64, Na)
 
     # If g-tensors are included in observables, they must be scalar. Precompute.
     gs = gs_as_scalar(swt, measure)
 
     for (iq, q) in enumerate(qpts.qs)
         q_global = cryst.recipvecs * q
-        q_reshaped = sys.crystal.recipvecs \ q_global
 
-        for branch in 1:3   # (q, q+k, q-k) modes for propagation wavevector k
+        for branch in 1:3   # (q-k, q, q+k) modes for propagation wavevector k
             energies = excitations!(T0, H, sswt, q; branch)
             view(disp, :, branch, iq) .= view(energies, 1:L)
             view(T, :, :, branch) .= T0
         end
 
         for i in 1:Na
+            r_global = global_position(sys, (1, 1, 1, i))
             ff = get_swt_formfactor(measure, 1, i)
-            c[i] = data.sqrtS[i] * gs[i] * compute_form_factor(ff, norm2(q_global))
-        end
-
-        for i in 1:L, j in 1:L
-            ui = R[i][:, 1] + im*R[i][:, 2]
-            uj = R[j][:, 1] + im*R[j][:, 2]
-            ri = sys.crystal.positions[i]
-            rj = sys.crystal.positions[j]
-            phase = exp(-2π * im*dot(q_reshaped, rj-ri))
-            for α in 1:3, β in 1:3
-                Y[i, j, α, β] = c[i] * c[j] * (ui[α] * conj(uj[β])) * phase
-                Z[i, j, α, β] = c[i] * c[j] * (ui[α] * uj[β]) * phase
-                V[i, j, α, β] = c[i] * c[j] * (conj(ui[α]) * conj(uj[β])) * phase
-                W[i, j, α, β] = c[i] * c[j] * (conj(ui[α]) * uj[β]) * phase
-            end
+            Avec_pref[i] = exp(- im * dot(q_global, r_global))
+            Avec_pref[i] *= compute_form_factor(ff, norm2(q_global))
         end
 
         for branch in 1:3, band in 1:L
-            t = view(T, :, band, branch)
-            for α in 1:3, β in 1:3
-                S[α, β, band, branch] = dot(t, view(YZVW, :, :, α, β), t) / 2Ncells
+            Avec = zero(CVec{3})
+            v = reshape(view(T, :, band, branch), Na, 2)
+            for i in 1:Na
+                R = local_rotations[i]
+                displacement = R * SA[v[i, 2] + v[i, 1], im * (v[i, 2] - v[i, 1]), 0.0]
+                Avec += Avec_pref[i] * (sqrtS[i]/sqrt(2)) * gs[i] * displacement
             end
+            S[band, branch] = Avec * Avec' / Ncells
         end
 
-        avg(S) = 1/2 * (S - nx * S * nx + (R2-I) * S * R2 + R2 * S * (R2-I) + R2 * S * R2)
-
         for band = 1:L
-            S[:, :, band, 1] .= avg(CMat3(view(S, :, :, band, 1))) * conj(R1)
-            S[:, :, band, 2] .= avg(CMat3(view(S, :, :, band, 2))) * R2
-            S[:, :, band, 3] .= avg(CMat3(view(S, :, :, band, 3))) * R1
+            case = spiral_propagation_case(k)
+            if case == 1
+                # The three branches (q-k, q, q+k) are equivalent when k = 0
+                # (modulo 1). Spread 1/3 of the intensity among every branch.
+                S[band, 1] *= 1/3
+                S[band, 2] *= 1/3
+                S[band, 3] *= 1/3
+                @assert S[band, 1] ≈ S[band, 2] ≈ S[band, 3]
+            elseif case == 2
+                S[band, 1] = R1 * S[band, 1] * R1 + conj(R1) * S[band, 1] * R1
+                S[band, 2] = R2 * S[band, 2] * R2
+                S[band, 3] = conj(R1) * S[band, 3] * conj(R1) +  R1 * S[band, 3] * conj(R1)
+            else
+                S[band, 1] = R1 * S[band, 1] * R1
+                S[band, 2] = R2 * S[band, 2] * R2
+                S[band, 3] = conj(R1) * S[band, 3] * conj(R1)
+            end
         end
 
         for branch in 1:3, band in 1:L
             corrbuf = map(measure.corr_pairs) do (α, β)
-                S[α, β, band, branch]
+                S[band, branch][α, β]
             end
             intensity[band, branch, iq] = thermal_prefactor(disp[band, branch, iq]; kT) * measure.combiner(q_global, corrbuf)
         end
